HoundConfig merges default and user configurations
Previously, `hound_config#enabled_for?` would return `true` for any
value that was not:

 - Explicitly set to false in the supplied config file
 - A language marked as beta

Additionally, HoundConfig didn't account for aliases. This could result in
someone disabling an alias (e.g. `javascript`) but still running the
actual linter (e.g. `jshint`). Instead of trying to perform boolean
logic about what is and is not activated, this strategy just starts with
a default configuration and merges the user-specified configuration into
it using a series of transforms.